move more common functions to helpers
