+++ crates/ruff_linter/src/rules/refurb/rules/write_whole_file.rs
@@ -2,9 +2,7 @@ use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, violation};
 use ruff_python_ast::visitor::{self, Visitor};
 use ruff_python_ast::{self as ast, Expr, Stmt};
-use ruff_python_codegen::Generator;
-use ruff_python_semantic::{BindingId, ResolvedReference, SemanticModel};
-use ruff_text_size::{Ranged, TextRange};
+use ruff_text_size::Ranged;
 
 use crate::checkers::ast::Checker;
 use crate::fix::snippet::SourceCodeSnippet;
@@ -58,7 +56,7 @@ pub(crate) fn write_whole_file(checker: &mut Checker, with: &ast::StmtWith) {
     }
 
     // First we go through all the items in the statement and find all `open` operations.
-    let candidates = find_file_opens(with, checker.semantic());
+    let candidates = helpers::find_file_opens(with, checker.semantic(), false);
     if candidates.is_empty() {
         return;
     }
@@ -77,138 +75,13 @@ pub(crate) fn write_whole_file(checker: &mut Checker, with: &ast::StmtWith) {
             Diagnostic::new(
                 WriteWholeFile {
                     filename: SourceCodeSnippet::from_str(&checker.generator().expr(open.filename)),
-                    suggestion: make_suggestion(open, content, checker.generator()),
+                    suggestion: helpers::make_suggestion(open, content, checker.generator()),
                 },
                 open.item.range(),
             )
         })
         .collect();
     checker.diagnostics.extend(diagnostics);
-}
-
-/// Find and return all `open` operations in the given `with` statement.
-fn find_file_opens<'a>(
-    with: &'a ast::StmtWith,
-    semantic: &'a SemanticModel<'a>,
-) -> Vec<helpers::FileOpen<'a>> {
-    with.items
-        .iter()
-        .filter_map(|item| find_file_open(item, with, semantic))
-        .collect()
-}
-
-/// Find `open` operation in the given `with` item.
-fn find_file_open<'a>(
-    item: &'a ast::WithItem,
-    with: &'a ast::StmtWith,
-    semantic: &'a SemanticModel<'a>,
-) -> Option<helpers::FileOpen<'a>> {
-    // We want to match `open(...) as var`.
-    let ast::ExprCall {
-        func,
-        arguments: ast::Arguments { args, keywords, .. },
-        ..
-    } = item.context_expr.as_call_expr()?;
-
-    if func.as_name_expr()?.id != "open" {
-        return None;
-    }
-
-    let var = item.optional_vars.as_deref()?.as_name_expr()?;
-
-    // Ignore calls with `*args` and `**kwargs`. In the exact case of `open(*filename, mode="w")`,
-    // it could be a match; but in all other cases, the call _could_ contain unsupported keyword
-    // arguments, like `buffering`.
-    if args.iter().any(Expr::is_starred_expr)
-        || keywords.iter().any(|keyword| keyword.arg.is_none())
-    {
-        return None;
-    }
-
-    // Match positional arguments, get filename and mode.
-    let (filename, pos_mode) = helpers::match_open_args(args)?;
-
-    // Match keyword arguments, get keyword arguments to forward and possibly mode.
-    let (keywords, kw_mode) = match_open_keywords(keywords)?;
-
-    let mode = kw_mode.unwrap_or(pos_mode);
-
-    if !matches!(
-        mode,
-        helpers::OpenMode::WriteText | helpers::OpenMode::WriteBytes,
-    ) {
-        return None;
-    }
-
-    // Path.write_bytes does not support any kwargs.
-    if matches!(mode, helpers::OpenMode::WriteBytes) && !keywords.is_empty() {
-        return None;
-    }
-
-    // Now we need to find what is this variable bound to...
-    let scope = semantic.current_scope();
-    let bindings: Vec<BindingId> = scope.get_all(var.id.as_str()).collect();
-
-    let binding = bindings
-        .iter()
-        .map(|x| semantic.binding(*x))
-        // We might have many bindings with the same name, but we only care
-        // for the one we are looking at right now.
-        .find(|binding| binding.range() == var.range())?;
-
-    // Since many references can share the same binding, we can limit our attention span
-    // exclusively to the body of the current `with` statement.
-    let references: Vec<&ResolvedReference> = binding
-        .references
-        .iter()
-        .map(|id| semantic.reference(*id))
-        .filter(|reference| with.range().contains_range(reference.range()))
-        .collect();
-
-    // And even with all these restrictions, if the file handle gets used not exactly once,
-    // it doesn't fit the bill.
-    let [reference] = references.as_slice() else {
-        return None;
-    };
-
-    Some(helpers::FileOpen {
-        item,
-        filename,
-        mode,
-        keywords,
-        reference,
-    })
-}
-
-/// Match keyword arguments. Return keyword arguments to forward and mode.
-fn match_open_keywords(
-    keywords: &[ast::Keyword],
-) -> Option<(Vec<&ast::Keyword>, Option<helpers::OpenMode>)> {
-    let mut result: Vec<&ast::Keyword> = vec![];
-    let mut mode: Option<helpers::OpenMode> = None;
-
-    for keyword in keywords {
-        match keyword.arg.as_ref()?.as_str() {
-            "encoding" | "errors" | "newline" => result.push(keyword),
-
-            // This might look bizarre, - why do we re-wrap this optional?
-            //
-            // The answer is quite simple, in the result of the current function
-            // mode being `None` is a possible and correct option meaning that there
-            // was NO "mode" keyword argument.
-            //
-            // The result of `match_open_mode` on the other hand is None
-            // in the cases when the mode is not compatible with `write_text`/`write_bytes`.
-            //
-            // So, here we return None from this whole function if the mode
-            // is incompatible.
-            "mode" => mode = Some(helpers::match_open_mode(&keyword.value)?),
-
-            // All other keywords cannot be directly forwarded.
-            _ => return None,
-        };
-    }
-    Some((result, mode))
 }
 
 /// AST visitor that matches `open` operations with the corresponding `write` calls.
@@ -285,27 +158,4 @@ fn match_write_call(expr: &Expr) -> Option<(&Expr, Vec<Expr>)> {
     }
 
     Some((attr.value.as_ref(), call.arguments.args.as_ref().to_vec()))
-}
-
-/// Construct the replacement suggestion call.
-fn make_suggestion(
-    open: &helpers::FileOpen<'_>,
-    write_arguments: Vec<Expr>,
-    generator: Generator,
-) -> SourceCodeSnippet {
-    let name = ast::ExprName {
-        id: open.mode.pathlib_method(),
-        ctx: ast::ExprContext::Load,
-        range: TextRange::default(),
-    };
-    let call = ast::ExprCall {
-        func: Box::new(name.into()),
-        arguments: ast::Arguments {
-            args: Box::from(write_arguments),
-            keywords: open.keywords.iter().copied().cloned().collect(),
-            range: TextRange::default(),
-        },
-        range: TextRange::default(),
-    };
-    SourceCodeSnippet::from_str(&generator.expr(&call.into()))
 }
